+++ dataservices/management/commands/tests/test_import_data.py
@@ -218,30 +218,59 @@ def test_import_worldbank_data_all(world_bank_mock):
 
 
 @pytest.fixture()
-def trade_by_quarter_raw_data():
+def trade_in_services_raw_data():
     return {
-        'ons_iso_alpha_2_code': ['CN', 'CN', 'CN', 'CN', 'CN', 'CN', 'XX'],
+        'iso2': ['CN', 'CN', 'CN', 'CN', 'CN', 'CN', 'XX'],
         'period': [
             'quarter/2021-Q1',
             'quarter/2021-Q2',
             'quarter/2021-Q3',
+            'quarter/2021-Q4',
+            'quarter/2022-Q1',
+            'year/2021',
+            'quarter/2021-Q1',
+        ],
+        'period_type': ['quarter', 'quarter', 'quarter', 'quarter', 'quarter', 'year', 'quarter'],
+        'service_code': [1, 1, 1, 1, 1, 1, 1],
+        'service_name': [
+            'Manufacturing',
+            'Manufacturing',
+            'Manufacturing',
+            'Manufacturing',
+            'Manufacturing',
+            'Manufacturing',
+            'Manufacturing',
+        ],
+        'imports': [1.0, 2.0, 3.0, 3.0, 2.0, 9.0, 1.0],
+        'exports': [1.0, 2.0, 3.0, 3.0, 2.0, 9.0, 1.0],
+    }
+
+
+@pytest.fixture()
+def trade_in_goods_by_quarter_raw_data():
+    return {
+        'iso2': ['CN', 'CN', 'CN', 'CN', 'CN', 'CN', 'XX'],
+        'period': [
             'quarter/2021-Q1',
             'quarter/2021-Q2',
             'quarter/2021-Q3',
+            'quarter/2022-Q1',
+            'quarter/2022-Q2',
+            'quarter/2022-Q3',
             'quarter/2021-Q1',
         ],
-        'direction': ['exports', 'exports', 'exports', 'imports', 'imports', 'imports', 'exports'],
-        'product_code': [1, 1, 1, 1, 1, 1, 1],
-        'product_name': [
-            'Manufacturing',
-            'Manufacturing',
-            'Manufacturing',
-            'Manufacturing',
-            'Manufacturing',
-            'Manufacturing',
-            'Manufacturing',
+        'commodity_code': [1, 1, 1, 1, 1, 1, 1],
+        'commodity_name': [
+            'Aircraft',
+            'Aircraft',
+            'Aircraft',
+            'Aircraft',
+            'Aircraft',
+            'Aircraft',
+            'Aircraft',
         ],
-        'value': [1.0, 2.0, 3.0, 3.0, 2.0, None, 1.0],
+        'imports': [1.0, 2.0, 3.0, 3.0, 2.0, None, 1.0],
+        'exports': [1.0, 2.0, 3.0, 3.0, 2.0, None, 1.0],
     }
 
 
@@ -251,9 +280,9 @@ def trade_by_quarter_raw_data():
 def test_import_uk_total_trade_data(read_sql_mock):
     mock_data = {
         'ons_iso_alpha_2_code': ['CN', 'CN', 'CN', 'CN', 'CN', 'CN', 'XX'],
-        'period': ['2021-Q1', '2021-Q2', '2021-Q3', '2021-Q1', '2021-Q2', '2021-Q3', '2021-Q1'],
-        'direction': ['exports', 'exports', 'exports', 'imports', 'imports', 'imports', 'exports'],
-        'value': [1.0, 2.0, 3.0, 3.0, 2.0, 1.0, 1.0],
+        'period': ['2021-Q1', '2021-Q2', '2021-Q3', '2021-Q4', '2022-Q1', '2022-Q2', '2021-Q1'],
+        'imports': [1.0, 2.0, 3.0, 3.0, 2.0, 1.0, 1.0],
+        'exports': [1.0, 2.0, 3.0, 3.0, 2.0, 1.0, 1.0],
     }
     read_sql_mock.return_value = [pd.DataFrame(mock_data)]
 
@@ -262,32 +291,32 @@ def test_import_uk_total_trade_data(read_sql_mock):
     management.call_command('import_countries')
     management.call_command('import_uk_total_trade_data')
 
-    assert len(models.UKTotalTradeByCountry.objects.all()) == 4
+    assert len(models.UKTotalTradeByCountry.objects.all()) == 7
 
 
 @pytest.mark.django_db
 @mock.patch('pandas.read_sql')
 @override_settings(DATA_WORKSPACE_DATASETS_URL='postgresql://')
-def test_import_uk_trade_in_services_data(read_sql_mock, trade_by_quarter_raw_data):
-    read_sql_mock.return_value = [pd.DataFrame(trade_by_quarter_raw_data)]
+def test_import_uk_trade_in_services_data(read_sql_mock, trade_in_services_raw_data):
+    read_sql_mock.return_value = [pd.DataFrame(trade_in_services_raw_data)]
 
     assert len(models.UKTradeInServicesByCountry.objects.all()) == 0
 
     management.call_command('import_countries')
     management.call_command('import_uk_trade_in_services_data')
 
-    assert len(models.UKTradeInServicesByCountry.objects.all()) == 3
+    assert len(models.UKTradeInServicesByCountry.objects.all()) == 6
 
 
 @pytest.mark.django_db
 @mock.patch('pandas.read_sql')
 @override_settings(DATA_WORKSPACE_DATASETS_URL='postgresql://')
-def test_import_uk_trade_in_goods_data(read_sql_mock, trade_by_quarter_raw_data):
-    read_sql_mock.return_value = [pd.DataFrame(trade_by_quarter_raw_data)]
+def test_import_uk_trade_in_goods_data(read_sql_mock, trade_in_goods_by_quarter_raw_data):
+    read_sql_mock.return_value = [pd.DataFrame(trade_in_goods_by_quarter_raw_data)]
 
     assert len(models.UKTradeInGoodsByCountry.objects.all()) == 0
 
     management.call_command('import_countries')
     management.call_command('import_uk_trade_in_goods_data')
 
-    assert len(models.UKTradeInGoodsByCountry.objects.all()) == 3
+    assert len(models.UKTradeInGoodsByCountry.objects.all()) == 6